// src/components/types.tsx
export interface UserData {
    firstName: string;
    lastName: string;
    proffesion: string;
    aboutSelf: string;
    school: string;
    degree: string;
    position: string;
    company: string;
    startOfEm: string;
    endOfEm: string;
    DailyTasks: string;
    address: string;
    eMail: string;
    phoneNum: string;
    LinkToP: string;
}